+++ src/client.cpp
@@ -66,13 +66,13 @@ void librg_cpp::Client::disconnect() {
     librg_network_stop(context());
 }
 
-void librg_cpp::Client::sendMessage(uint32_t id, void *data, size_t size) {
+void librg_cpp::Client::sendMessage(uint16_t id, void *data, size_t size) {
     assert(data != nullptr);
 
     sendMessageToAll(id, data, size);
 }
 
-void librg_cpp::Client::sendMessage(uint32_t id, const std::shared_ptr<Data> &data) {
+void librg_cpp::Client::sendMessage(uint16_t id, const std::shared_ptr<Data> &data) {
     assert(data != nullptr);
 
     sendMessageToAll(id, data->raw(), data->writePosition());
